Fixed FPV footer not displaying information
