--- Assets/UI Toolkit/Scripts/Behaviours/FooterBehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f3adb988a4a3c4e4ea91fac29b0a7601
