--- Assets/_Application/Cameras/CameraService/Runtime/CameraPositionWatcher.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7489ce4d5fba4446bcbed7dd772d46ba
-timeCreated: 1773300220
