+++ Assets/_Application/Cameras/CameraService/Runtime/CameraService.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 8564524906190026714}
   - component: {fileID: 7470763439909778047}
   - component: {fileID: 6938902015790729898}
-  - component: {fileID: 7274287954263651405}
   m_Layer: 0
   m_Name: CameraService
   m_TagString: Untagged
@@ -65,30 +64,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   service: {fileID: 7470763439909778047}
---- !u!114 &7274287954263651405
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5024820877031905233}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7489ce4d5fba4446bcbed7dd772d46ba, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: eu.netherlands3d.twin.cameras::Netherlands3D.Twin.Cameras.CameraPositionWatcher
-  OnPositionChanged:
-    m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 11400000, guid: d4a74b03cccc58345974ac5fcbc59d0c, type: 2}
-        m_TargetAssemblyTypeName: Netherlands3D.Events.Vector3Event, eu.netherlands3d.event-system.Runtime
-        m_MethodName: InvokeStarted
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
