+++ Assets/_Application/Services/LayerMessageService.cs
@@ -1,6 +1,8 @@
 using Netherlands3D.DataTypeAdapters;
+using Netherlands3D.Events;
 using Netherlands3D.Twin.Layers;
 using UnityEngine;
+using UnityEngine.Events;
 
 namespace Netherlands3D.Twin.Services
 {
@@ -12,7 +14,9 @@ namespace Netherlands3D.Twin.Services
         private int activeCounter;
         private DataTypeChain[] chains;
         
-
+        [SerializeField] private StringEvent layerSourceAttributionEvent;
+        public UnityEvent<string> OnAttributionReceived;
+        
         private void Awake()
         {
             layers = App.Layers;
@@ -26,6 +30,7 @@ namespace Netherlands3D.Twin.Services
             layers.LayerAdded.AddListener(OnLayerAdded);
             layers.LayerRemoved.AddListener(OnLayerRemoved);
             layers.VisualizationCreated.AddListener(OnVisualizationCreated); // when the visualisation is created, we want to listen to potential error messages (eg. parse errors) to display
+            layerSourceAttributionEvent.AddListenerStarted(OnAttributionReceived.Invoke);
 
             foreach (var chain in chains)
             {
@@ -41,6 +46,7 @@ namespace Netherlands3D.Twin.Services
             layers.LayerAdded.RemoveListener(OnLayerAdded);
             layers.LayerRemoved.AddListener(OnLayerRemoved);
             layers.VisualizationCreated.RemoveListener(OnVisualizationCreated);
+            layerSourceAttributionEvent.RemoveListenerStarted(OnAttributionReceived.Invoke);
             
             foreach (var chain in chains)
             {
